Implemented the new input system (still problematic)
I was able to change from the old input sytem to the new one. I added one Action map (Player), with movement and look. the code and setting have been added like they did in class. The outcomes of the program don't align with what i expected which was a moving cube at this point, but the cube is not doing that, its spinning i think I am missing something. Still Working on it.