Changed position of the player and enemy
I placed the enemy at the edge of the roof and player near the rooftop entrance. The problem for now is that the enemy does not case the player and the patrol function is not functioning well, the enemy does not patrol as described in the script ( it does not move at all, it only rotates to angle of the player and keeps on shooting without moving.