Trying to add throw mechanic (Error)
Wrote the whole code, its giving me an error on the calling throw function add the end of the code where it tells me that my pickup object script does not contain the function "throw". Still need to figure this out. If it keeps on giving me issues i shall focus on the Winning condition and Health system so that the game gains a little bit of life.