Update Enemy health enemy gets destroyed
The enemy gets destroyed after its life has ended.

Now i need to implement colliders on all the enemies so that this works. On the enemies on the rooftop.... Also i need to implement the shooting system on the enemies too so that they can be able to kill the player. Then that will enable us to add a winning condition coz when the player is destroyed nothing can be done.

Les go!!!