Mathew's Advice (Scroller for the health, the stairs, the materials,  )
I still need to figure out, how to implement what I and Mathew were working on in class. The scroller. The materials (need to be applied on both sides of the model).  The stairs (The gaps should not be present because they mess around with the PLAYER CONTROLLER component) specifically the player offset which determines where the player steps. Mat, also suggested we use multiple box colliders on the stairs (maybe rotate them a bit to give the player some assistance ( Box collider one =/ ; Box Collider 2 \...... etc)

- We need to add sound ( We can use a slider to bring the volume down and up)

-We can create the sound using FL Studio, original sound.

- Also, need to adjust the mechanics.  The interact mechanic need a distance variable because it take the player a while to interact with the Object.

       Some things to consider
-Pickables (bringing the game to life)-

-bullets
-healables
-notes
- Grenades (Also particle systems), Smoke (Inspired by Dube's Fog idea), normal grenades

UI
- bullets
- Weapons

-Other things-

- Animation (Rigging the characters, animating the environment)

- ---- might forgot some --