PickUp and throw mechanic work
The pickUp mechanic has been changing shape for a while now. Now it no longer does that. Any object that you pick up you can actually throw it or just drop it. Moving on the the GUN.

I will focus on why the bullets aint strong and the particle system.

we need a aim mechanic....